DocTr/requirements.txt
python-doctr[torch]
pillow
pymupdf